Selector: SystemC module to create configuration parts from a conf' section
This SystemC module creates configuration parts from a bit vector
section of an arbitrary length. During creation of a instance of this
class, the start position of the selection as well as its length are
defined. From this selection an arbitrary set of parts with a
parameterizable bit width are created as output. This class can bit
used to select a part of a bit vector configuration and connect this
parts to its targets.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>