Synchronizer: Used to show the end of processing to VCGRA caller
This module is placed at the end of a VCGRA design to collect the valid
signals of all last PEs. Together with a confiuration bitstream where
all unused PE valid signals are overwritten by '1', this module raises
an ready signal to high for one clock cycle if all used PEs of last level
show a valid content.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>