MMU: Update design to handle block transmissions
Add new cases in state machine to decode block or single transmissions by
place port input value. If a single value should be transmitted, the place
value is validated to fit to selected cache. If not, no transaction is done
and a Error message is written. Another case handles address and place
adaptions for every transmission until cache is filled.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>