TB_DataInCache: Add test bench for data input cache
The design tests a small cache with 2 lines and 4 places each. Data is
written into the cache and the configuration is switched. A couple of
limit checks is tested, others are prepared but outcommented, because the
design was to small to test it (number if bits for line and place was to
small to evaluate differences).

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>