DataOutCache: Improve dump function and correct error behaviour
Changing current input cache line and updating line with two functions did
not work with pointer or refernce variable. Thus, we store the current
cache line as a value in line switching function and load the cache line
in a temporary reference in update function. A cache line update for
input values is only processed, if the internal stored index and the new
selection index differ. Otherwise a warning occured every clock cycle.

If input and output cache line differ, a cache line can be loaded while the
other one is being updated now.

Add and correct dump information.