TB_DataOutCache: Add test bench for data output cache
The test bench creates a set of result values for data cache inputs and
trigger the update of the current selected input cache line. Also, a
cache line is loaded by addressing an output cache line and iterating over
its places. Update and value transfer is done sequentually or in parallel.

A couple of possible error states are tested.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>