TB_MMU and sc_main: Add Testbench for MMU
This branch is used to test the MMU with shared memory and all it's
connected caches. First the data in cache is targeted to get some values.
We transmit every value, each as a single command for the MMU. Afterwards
we change the cache line of the data input cache to show the new values.
Then we transmit a block of values to the data input cache.
Afterwards we set new result values at the data out cache and update a
cache line in data output cache. This cache line is than transmitted to
shared memory as a block.
Then, we load the PE and the virtual channel caches with configuration
streams, stored in shared memory. This branch is used to test the MMU
with shared memory

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>