sc_main: Update sc_main for ManagementUnit unit test
This small example tests a ManagementUnit instance with its corresponding
test bench. The starts, pauses and resets the ManagementUnit. If access
to memory or VCGRA is performed, the test bench simulates these components
by simply reacting with the ready signals after a programmed timeout.
The tests traces all signals between ManagementUnit and test bench as well
as the internal state variable.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>