Update sc_main.cpp for new VCGRA architecture
This version of the sc_main.cpp includes the full architecture witch
caches, processing system MMU and management unit. The architecture is
controlled by assembler commands. Data is loaded from memory into the
data caches of the design. Afterwards, data is processed and stored in
the data output caches. At the end, we store the results back into memory.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>