MMU: Correct typo in doxygen docu and adaptions in typedefs
Cache select type is shifted to Typedef.h, because it may be used
in different components.
Bitwidth for cache_place_type_t is calculated in Typedef.h. Because
the number of inputs and outputs may differ, the necessary bitwidth is
calculated by the maximal number of places within the data caches. Thus,
the signal can be shared for both DataInCache and DataOutCache.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>