Management_Unit: Correct issue for wait_ready interrupt
The logic had a failure where the response of the VCGRA was not recognized,
if the MMU was used when the interrupt occurred. This is fixed changed by
using a FIFO instead of a simple signal. With non-blocking reads and writes
the state machine only acts if a value can be read from the FIFO. Otherwise
it continues waiting without changing the state. The FIFO has only one
place that no multiple interrupts are possible to process.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>