mcpat: Update router configuration vor virtual channel model
Each NoC consists of routers where each router represents a MUX. So
the router has alsways on output and inputs as number of inputs of the
virtual channel. The number of virtical nodes is set to 1 and the number
of horizontal nodes is set to the number of channel outputs.

In FPGA design 3% of the chip area is vor channels. In McPAT with this
model, 4% of the chip area is for channels.

Signed-off-by: andrewerner <andre.werner-w2m@ruhr-uni-bochum.de>